* ReadPipe returns ssize_t

+++ src/con_dosx.cpp
@@ -1054,9 +1054,9 @@ int GUI::SetPipeView(int id, EModel *notify) {
     return 0;
 }
 
-int GUI::ReadPipe(int id, void *buffer, int len) {
+ssize_t GUI::ReadPipe(int id, void *buffer, int len) {
 #ifdef DJGPP
-    int rc;
+    ssize_t rc;
     if (id < 0 || id >= MAX_PIPES)
         return -1;
     if (Pipes[id].used == 0)
