+++ src/con_nt.cpp
@@ -1521,7 +1521,7 @@ int GUI::SetPipeView(int id, EModel *notify) /*FOLD00*/
 }
 
 
-int GUI::ReadPipe(int id, void *buffer, int len) /*FOLD00*/
+ssize_t GUI::ReadPipe(int id, void *buffer, int len) /*FOLD00*/
 {
     //int     l;
     GPipe*  p;
@@ -1813,8 +1813,8 @@ int GUI::SetPipeView(int id, EModel *notify) {
     return 0;
 }
 
-int GUI::ReadPipe(int id, void *buffer, int len) {
-    int l;
+ssize_t GUI::ReadPipe(int id, void *buffer, int len) {
+    ssize_t l;
     //ULONG ulPostCount;
 
     if (id < 0 || id > MAX_PIPES)
